core/standards/fullscreen/interactive/layout-fullscreen-selector.html

<!DOCTYPE html>
<html>
<head>
	<title>:fullscreen selector matching</title>
	<meta name="assert" content="The :fullscreen pseudo-class must match the top
	element of the context object's fullscreen element stack (if any).">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<style>
		:fullscreen {
			background-color: rgb(0, 200, 0);
		}
	</style>
	<script>
		var test = async_test(":fullscreen selector matching");

		document.addEventListener('fullscreenchange', function(e)
		{
			test.step(function() {
				var el = document.fullscreenElement;
				assert_equals(getComputedStyle(el).backgroundColor, 'rgb(0, 200, 0)', "element's background color");
			});

			test.done();
			removeFullscreenElement();
		}, false);

		function removeFullscreenElement()
		{
			var el = document.fullscreenElement;
			el.parentNode.removeChild(el);
		}
	</script>
</head>
<body>

	<p>Fullscreen element should have green background.</p>

	<button id="fs" onclick="this.requestFullscreen();">enter fullscreen</button>

	<div id="log"></div>

</body>
</html>
